requirements.txt
google-genai
python-dotenv
pillow
